Rename reducer functions to their state counterparts
