+++ client/reducers/symbols.js
