+++ client/reducers/symbolsOrder.js
